Update hubs sdk api
