+++ .babelrc
@@ -33,6 +33,7 @@
     "@babel/proposal-object-rest-spread",
     // Samsung Internet on the Oculus Go version is stuck at version 5.2, which is a 
     // Chromium 51, as of this writing. It needs babel to transpile async/await.
-    "@babel/plugin-transform-async-to-generator"
+    "@babel/plugin-transform-async-to-generator",
+    "@babel/plugin-proposal-export-namespace-from"
   ]
 }
