+++ example-plugin/pages/index.js
@@ -6,8 +6,8 @@ const {
   config,
   React: {
     Common: { Page, IfFeature, AuthContext },
-    Media: { Tiles: MediaTiles, Styles: mediaBrowserStyles },
-    HomePage: { PWAButton, CreateRoomButton, useFeaturedRooms, useHomePageRedirect, Styles: styles, discordLogoSmall }
+    Media: { Tiles: MediaTiles, styles: mediaBrowserStyles },
+    HomePage: { PWAButton, CreateRoomButton, useFeaturedRooms, useHomePageRedirect, styles, discordLogoSmall }
   }
 } = Hubs;
 import customStyles from "./index.css";
