+++ hubs-client/package.json
@@ -1,6 +1,6 @@
 {
   "name": "hubs-client",
-  "version": "0.0.5",
+  "version": "0.0.6",
   "description": "Pre-built Hubs client",
   "license": "MPL-2.0",
   "homepage": "https://github.com/mozilla/hubs#readme",
