+++ src/index.js
@@ -10,7 +10,7 @@ import { lang, messages } from "./utils/i18n";
 import { AuthContextProvider } from "./react-components/auth/AuthContext";
 import configs from "./utils/configs";
 import en from "react-intl/locale-data/en";
-import { initSDKContext } from "./sdk/landing-page-globals";
+import { initSDKContext } from "./sdk/home-page";
 
 addLocaleData([...en]);
 registerTelemetry("/home", "Hubs Home Page");
@@ -37,6 +37,9 @@ async function main() {
 
   if (configs.hasPlugin("home-page")) {
     initSDKContext(store);
+
+    console.log(window.Hubs);
+
     const plugin = await configs.importPlugin("home-page");
 
     if (plugin.HomePage) {
