+++ src/sdk/react/common.js
@@ -1,0 +1,6 @@
+export { default as IfFeature } from "../../react-components/if-feature";
+export { Page } from "../../react-components/layout/Page";
+export { Header } from "../../react-components/layout/Header";
+export { Footer } from "../../react-components/layout/Footer";
+export { default as styles } from "../../react-components/layout/Page.scss";
+export { AuthContext } from "../../react-components/auth/AuthContext";
