+++ src/sdk/react/index.js
@@ -1,0 +1,11 @@
+import React from "react";
+import ReactDOM from "react-dom";
+import PropTypes from "prop-types";
+import * as ReactIntl from "react-intl";
+import ClassNames from "classnames";
+
+window.React = React;
+window.ReactDOM = ReactDOM;
+window.ReactIntl = ReactIntl;
+window.PropTypes = PropTypes;
+window.ClassNames = ClassNames;
