+++ src/sdk/react/media.js
@@ -1,0 +1,2 @@
+export { default as Tiles } from "../../react-components/media-tiles";
+export { default as styles } from "../../assets/stylesheets/media-browser.scss";
